Added validation. Fixed formula for routes

+++ PECourseProject/Point.h
@@ -5,9 +5,9 @@ class Point
 public:
 	// This constructor should be used only for creating dummy objects which will be populated via input stream, for everything else use the parametized constructor
 	Point();
-	Point(float x, float y);
+	Point(double x, double y);
 
-	float distanceTo(Point other);
+	double distanceTo(Point other);
 
 	friend std::ostream& operator<<(std::ostream& os, const Point& point);
 	friend std::istream& operator>>(std::istream& os, Point& point);
